Relax compiler requirements at runtime
The runtime dependencies had compiler('cxx'). The intention here was to
make sure that some C++ compiler is available. However, this actually
forces this to be the compiler that was used when building the package,
i.e., GCC 9. This causes trouble when trying to install a package that
requires an older compiler at the same time, e.g., when installing
cppyy=2.3.0 and pythran=0.11.0. (see
https://github.com/conda-forge/pythran-feedstock/pull/65)